requirements.txt
arrow==1.2.2
asgiref==3.5.2
attrs==21.4.0
base58==0.2.2
click==6.6
coreapi==2.3.3
coreschema==0.0.4
CurrencyConverter==0.14.1
Django==3.1.14
django-extensions==3.1.5
djangorestframework==3.13.1
docker-py==1.8.0
drf-yasg==1.20.0
ecdsa==0.17.0
entrypoints==0.3
flake8==3.7.9
funcsigs==1.0.2
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.3
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.16.0
marshmallow-dataclass==8.5.8
mccabe==0.6.1
mnemonic==0.13
more-itertools==8.2.0
mypy-extensions==0.4.3
packaging==20.3
path==13.1.0
path.py==12.4.0
pbkdf2==1.3
pexpect==4.8.0
pika==1.1.0
pluggy==0.13.1
protobuf==3.0.0a3
psycopg2-binary==2.9.3
ptyprocess==0.6.0
py==1.10.0
pyaes==1.6.1
pycodestyle==2.5.0
pycryptodome==3.14.1
pyflakes==2.1.1
pyparsing==2.4.6
Pypubsub==4.0.3
pyrsistent==0.15.7
pytest==5.4.1
python-bitcoinrpc==1.0
python-dateutil==2.8.1
pytz==2019.3
pywallet==0.1.0
PyYAML==6.0
requests==2.11.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sha256==0.3
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.3.1
tabulate==0.8.6
two1==3.10.9
typing-inspect==0.7.1
typing_extensions==4.2.0
uritemplate==3.0.1
wcwidth==0.1.8
websocket-client==0.57.0
zipp==3.1.0
